Move wrapContract from utils to eth/smartContract
